put class in header file
